+++ CHANGELOG.md
@@ -8,7 +8,9 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ### Added
 
-- Added `sr_cement` operation type in `tezos.operations` index to process Smart Rollup Cemented Commitments.
+- config: Publish JSON schemas for config validation and autocompletion.
+- starknet.node: Added Starknet node datasource for last mile indexing.
+- tezos.operations: Added `sr_cement` operation type to process Smart Rollup Cemented Commitments.
 
 ### Fixed
 
@@ -44,14 +46,14 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 - config: Allow `sentry.dsn` to be empty string.
 - models: Fixed setting default value for `Meta.maxsize`.
-- starknet.subsquid.events: Fixed filtering events by key.
+- starknet.events: Fixed filtering events by key.
 
 ## [8.0.0b1] - 2024-06-19
 
 ### Added
 
 - cli: Added full project migration support for 3.0 spec.
-- starknet.subsquid.events: Added `starknet.subsquid.events` index kind to process Starknet events.
+- starknet.events: Added `starknet.events` index kind to process Starknet events.
 - starknet.subsquid: Added `starknet.subsquid` datasource to fetch historical data from Subsquid Archives.
 
 ### Fixed